feat: datastar port (#12)
* feat: start adding datastar to project

The SDK is being ported to JS from the TypeScript SDK

* refactor: upcase component function name format

* fix: correct typos and incorrect port

* chore: remove htmx and add datastar and sdk to sw

* feat: start porting some template to datastar

* chore: rename confirmation-handler component

* feat: add event mechanism for confirmation handler

* refactor: template index page on the backend

* refactor: add datastar redirect support

* refactor: move datastar handlers to correct response

The markup is now the only thing pending to be
re-written with datastar rather than HTMX.

* chore: remove dead code

* fix: revert edit to anchor element

* refactor: TodoList always returns a string

There is still a bug with the filter state
implementation. The assertions are failing.

* chore: add todo to handle complete on filtered

The filter bug is no longer there. I might
have forgotten to use the latest service worker.

* feat: add filtered update on complete

This is a little excessive on purpose to demo
the capability of streaming updates to client
so there is some duplication of markup
generation.

* chore: formatting and minor refactoring

* refactor: convert some Typescript types to JSDocs

I'm not sure how to type some of this with
JSDocs. Might need to resort to .d.ts files.